Updated comment system
It now uses a query instead of JSON, which should not only
be more lightweight but also work with browsers that do not
support Javascript.

Config has also been updated to allow loading files directly.
This change will be needed in the future for config reloads.

It is now also again possible to remove redirects from the
command line.

There are also a few other changes, but those aren't quite
noteworthy.